[refactor] LookTab componenet로 분리
