+++ src/constant/lookTab.ts
@@ -1,0 +1,6 @@
+export const lookTabCategory = [
+  { text: '베스트룩 👍', href: '/', id: 1 },
+  { text: '기온별 추천룩 🌤', href: '/temperatures', id: 2 },
+  { text: 'TPO 추천룩 👔', href: '/tpo', id: 3 },
+  { text: '활동성 추천룩 🏄', href: '/activity', id: 4 },
+];
