+++ src/types/Common/index.ts
@@ -25,3 +25,8 @@ export interface IFiltersProps {
 }
 
 export interface IBottomFilterProps extends IFiltersProps {}
+
+export interface ILookTabItemProps {
+  text: string;
+  href: string;
+}
